// lib/graphql/generated/graphql.ts
export type InputMaybe<T> = T | null | undefined;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** An ISO 8601-encoded date */
  ISO8601Date: { input: any; output: any; }
  /** An ISO 8601-encoded datetime */
  ISO8601DateTime: { input: any; output: any; }
  Upload: { input: any; output: any; }
  /** A valid URL, transported as a string */
  Url: { input: any; output: any; }
};
// ...
export type Category = {
  __typename?: 'Category';
  createdAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  fullPath: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  name?: Maybe<Scalars['String']['output']>;
  rankable: Scalars['Boolean']['output'];
  updatedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  videos?: Maybe<Array<OriginalVideo>>;
};
// ...
export type Creator = {
  __typename?: 'Creator';
  avatar?: Maybe<Scalars['String']['output']>;
  createdAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  id: Scalars['ID']['output'];
  name?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
};
// ...
export type Duration = {
  __typename?: 'Duration';
  minuteUnitName?: Maybe<Scalars['String']['output']>;
  minutes?: Maybe<Scalars['Int']['output']>;
  secondUnitName?: Maybe<Scalars['String']['output']>;
  seconds?: Maybe<Scalars['Int']['output']>;
};
// ...
export type OriginalVideo = Video & {
  __typename?: 'OriginalVideo';
  categories?: Maybe<Array<Category>>;
  comment?: Maybe<Comment>;
  commentaryUrl?: Maybe<Scalars['String']['output']>;
  comments?: Maybe<CommentConnection>;
  commentsCount: Scalars['Int']['output'];
  contentEnd?: Maybe<Duration>;
  cookies?: Maybe<Cookies>;
  createdAt: Scalars['ISO8601DateTime']['output'];
  creator?: Maybe<Creator>;
  currentUserHasLike?: Maybe<Scalars['Boolean']['output']>;
  currentUserVideoRating?: Maybe<VideoRating>;
  description?: Maybe<Scalars['String']['output']>;
  drmDashFileUrl?: Maybe<Scalars['String']['output']>;
  drmHlsFileUrl?: Maybe<Scalars['String']['output']>;
  duration: Duration;
  fileUrl?: Maybe<Scalars['String']['output']>;
  filmarksUrl?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  jiqooNextVideo?: Maybe<OriginalVideo>;
  landscapeThumbnail?: Maybe<Scalars['String']['output']>;
  likeNum: Scalars['Int']['output'];
  nextVideo?: Maybe<OriginalVideo>;
  playlists?: Maybe<Array<Playlist>>;
  playlistsVideos?: Maybe<Array<PlaylistsVideo>>;
  portraitThumbnail?: Maybe<Scalars['String']['output']>;
  publishedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  recommends?: Maybe<Array<OriginalVideo>>;
  satisfiedNum: Scalars['Int']['output'];
  status: VideoStatusEnum;
  subtitles?: Maybe<Array<SrtSubtitle>>;
  thumbnailCategory?: Maybe<Category>;
  title?: Maybe<Scalars['String']['output']>;
  trailer?: Maybe<TrailerVideo>;
  translations: Array<VideoTranslation>;
  updatedAt: Scalars['ISO8601DateTime']['output'];
  viewingTime?: Maybe<VideoViewingTime>;
  vtts?: Maybe<Array<VttSubtitle>>;
  willBePublishedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  willBeUnpublishedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
};
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars['String']['output']>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars['Boolean']['output'];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars['Boolean']['output'];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars['String']['output']>;
};
// ...
export type Query = {
  __typename?: 'Query';
  adminCoupons: CouponConnection;
  adminImages: AdminImageConnection;
  adminOriginalVideos: OriginalVideoConnection;
  adminScreenPopup?: Maybe<ScreenPopup>;
  adminScreenPopups: ScreenPopupConnection;
  adminTestMovies: Array<TestMovie>;
  adminUsers: UserConnection;
  allOriginalVideos: Array<OriginalVideo>;
  category: Category;
  commentReplies: CommentRepliesConnection;
  coupon?: Maybe<Coupon>;
  currentUser?: Maybe<User>;
  firstRecommendationRow: RecommendationRow;
  homeScreens: Array<HomeScreen>;
  jiqooCategory: Category;
  jiqooCommentReplies: CommentRepliesConnection;
  jiqooHomeScreens: Array<HomeScreen>;
  jiqooOriginalVideo?: Maybe<OriginalVideo>;
  jiqooRootCategories: Array<RootCategory>;
  jiqooScreenPopups: ScreenPopupConnection;
  jiqooTrailerVideo?: Maybe<TrailerVideo>;
  jiqooTrailerVideos: TrailerVideoConnection;
  jiqooVideoComments: VideoCommentsConnection;
  jiqooVideoRecommends?: Maybe<VideoRecommendsConnection>;
  notifications: UserNotificationConnection;
  opsToggles: OpsTogglesResult;
  originalVideo?: Maybe<OriginalVideo>;
  payment: Payment;
  playlist: Playlist;
  playlists: Array<Playlist>;
  promotionCode?: Maybe<PromotionCode>;
  recommendedSearchKeywords: Array<RecommendedSearchKeyword>;
  rootCategories: Array<RootCategory>;
  screenPopups: ScreenPopupConnection;
  trailerVideo?: Maybe<TrailerVideo>;
  trailerVideos: TrailerVideoConnection;
  trialVideos: Array<OriginalVideo>;
  tvSearchCategories: Array<TvSearchCategory>;
  userNotificationSettings: Array<UserNotificationSetting>;
  videoAiRecommends?: Maybe<VideoRecommendsConnection>;
  videoComments: VideoCommentsConnection;
  videoHistories: VideoHistoryConnection;
  videoRatings: VideoRatingConnection;
  videoRecommends?: Maybe<VideoRecommendsConnection>;
  withdrawalVideos: Array<TrailerVideo>;
};
// ...
export type VideoCommentsConnection = {
  __typename?: 'VideoCommentsConnection';
  allCount?: Maybe<Scalars['Int']['output']>;
  /** A list of edges. */
  edges?: Maybe<Array<VideoCommentsEdge>>;
  id: Scalars['ID']['output'];
  /** A list of nodes. */
  nodes?: Maybe<Array<Comment>>;
  /** Information to aid in pagination. */
  pageInfo?: Maybe<PageInfo>;
  totalCount?: Maybe<Scalars['Int']['output']>;
};
// ...
export type GetOriginalVideoQueryVariables = Exact<{
  id: Scalars['ID']['input'];
}>;


export type GetOriginalVideoQuery = { __typename?: 'Query', originalVideo?: { __typename?: 'OriginalVideo', title?: string | null, description?: string | null, landscapeThumbnail?: string | null, likeNum: number, duration: { __typename?: 'Duration', minutes?: number | null, seconds?: number | null }, creator?: { __typename?: 'Creator', id: string, name?: string | null, avatar?: string | null } | null, categories?: Array<{ __typename?: 'Category', id: string, name?: string | null }> | null } | null };

export type GetVideoCommentsQueryVariables = Exact<{
  id: Scalars['ID']['input'];
  first?: InputMaybe<Scalars['Int']['input']>;
  after?: InputMaybe<Scalars['String']['input']>;
}>;


export type GetVideoCommentsQuery = { __typename?: 'Query', videoComments: { __typename?: 'VideoCommentsConnection', id: string, allCount?: number | null, pageInfo?: { __typename?: 'PageInfo', startCursor?: string | null, endCursor?: string | null, hasNextPage: boolean, hasPreviousPage: boolean } | null, edges?: Array<{ __typename?: 'VideoCommentsEdge', cursor: string, node?: { __typename?: 'Comment', id: string, contents?: string | null, createdAt?: any | null, likeNum?: number | null, user?: { __typename?: 'User', id: string, name?: string | null, avatar?: string | null } | null } | null }> | null } };
// ...
export const GetOriginalVideoDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"getOriginalVideo"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"id"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"ID"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"originalVideo"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"id"},"value":{"kind":"Variable","name":{"kind":"Name","value":"id"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"title"}},{"kind":"Field","name":{"kind":"Name","value":"description"}},{"kind":"Field","name":{"kind":"Name","value":"landscapeThumbnail"}},{"kind":"Field","name":{"kind":"Name","value":"likeNum"}},{"kind":"Field","name":{"kind":"Name","value":"duration"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"minutes"}},{"kind":"Field","name":{"kind":"Name","value":"seconds"}}]}},{"kind":"Field","name":{"kind":"Name","value":"creator"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"avatar"}}]}},{"kind":"Field","name":{"kind":"Name","value":"categories"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}}]}}]}}]}}]} as unknown as DocumentNode<GetOriginalVideoQuery, GetOriginalVideoQueryVariables>;
export const GetVideoCommentsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"getVideoComments"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"id"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"ID"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"first"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"Int"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"after"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"videoComments"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"id"},"value":{"kind":"Variable","name":{"kind":"Name","value":"id"}}},{"kind":"Argument","name":{"kind":"Name","value":"first"},"value":{"kind":"Variable","name":{"kind":"Name","value":"first"}}},{"kind":"Argument","name":{"kind":"Name","value":"after"},"value":{"kind":"Variable","name":{"kind":"Name","value":"after"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"allCount"}},{"kind":"Field","name":{"kind":"Name","value":"pageInfo"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"startCursor"}},{"kind":"Field","name":{"kind":"Name","value":"endCursor"}},{"kind":"Field","name":{"kind":"Name","value":"hasNextPage"}},{"kind":"Field","name":{"kind":"Name","value":"hasPreviousPage"}}]}},{"kind":"Field","name":{"kind":"Name","value":"edges"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"cursor"}},{"kind":"Field","name":{"kind":"Name","value":"node"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"contents"}},{"kind":"Field","name":{"kind":"Name","value":"user"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"avatar"}}]}},{"kind":"Field","name":{"kind":"Name","value":"createdAt"}},{"kind":"Field","name":{"kind":"Name","value":"likeNum"}}]}}]}}]}}]}}]} as unknown as DocumentNode<GetVideoCommentsQuery, GetVideoCommentsQueryVariables>;